Move React components to dir
