+++ js/test/header.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { expect } from 'chai';
 
-import Header from '../header';
+import Header from 'components/header';
 
 describe('Header', () => {
   let wrapper;
