+++ webpack.config.js
@@ -23,5 +23,6 @@ module.exports = {
   },
   resolve: {
     extensions: ['.js', '.jsx', '.json'],
+    modules: [path.join(__dirname, 'js'), 'node_modules'],
   },
 };
