+++ webpack.config.test.js
@@ -18,5 +18,6 @@ module.exports = {
   },
   resolve: {
     extensions: ['.js', '.jsx', '.json'],
+    modules: [path.join(__dirname, 'js'), 'node_modules'],
   },
 };
